🎨 improves: code style

+++ src/main/java/io/th0rgal/oraxen/utils/Utils.java
@@ -18,7 +18,7 @@ public class Utils {
                 (BlockFace.EAST, BlockFace.WEST, BlockFace.SOUTH, BlockFace.NORTH, BlockFace.DOWN, BlockFace.UP);
         int sum = 0;
         for (BlockFace blockFace : blockData.getFaces())
-            sum += Math.pow(2, properties.indexOf(blockFace));
+            sum += (int)Math.pow(2, properties.indexOf(blockFace));
         return sum;
     }
